do not open the user settings this much! merge_user_settings are invoked four or five times. Although the documentation from load_settings says: "Subsequent calls to load_settings() with the base_name will return the same object, and not load the settings from disk again." I feel that it's bad practice to execute the function merge_user_settings multiple times.
So here is an overhaul:
renamed the function: `merge_user_settings` to `get_user_settings`.
only where it is needed the settings are loaded with settings=get_user_settings()
then when changing or getting some settings, no new function call is executed with `merge_user_settings` as it was before, but
it only access the already loaded settings.
And it won't pass a dictionary of transformation to the pandoc command anymore, but only the name of the transformation and
then only after that it will look the items from the picked transformation.